+++ Wire-iOS/Sources/LaunchSequenceOperation.swift
@@ -19,6 +19,9 @@
 import Foundation
 import WireSyncEngine
 import WireCommonComponents
+import AppCenter
+import AppCenterCrashes
+import AppCenterDistribute
 import avs
 
 // MARK: - LaunchSequenceOperation
@@ -98,3 +101,88 @@ final class FileBackupExcluderOperation: LaunchSequenceOperation {
     }
 }
 
+// MARK: - AppCenterOperation
+final class AppCenterOperation: NSObject, LaunchSequenceOperation {
+    private var zmLog: ZMSLog {
+        return ZMSLog(tag: "UI")
+    }
+    
+    public func execute() {
+        guard AutomationHelper.sharedHelper.useAppCenter || Bundle.useAppCenter else {
+            MSAppCenter.setTrackingEnabled(false)
+            return
+        }
+        UserDefaults.standard.set(true, forKey: "kBITExcludeApplicationSupportFromBackup") //check
+        
+        guard !TrackingManager.shared.disableCrashSharing else {
+            MSAppCenter.setTrackingEnabled(false)
+            return
+        }
+        
+        MSCrashes.setDelegate(self)
+        MSDistribute.setDelegate(self)
+        
+        MSAppCenter.start()
+        
+        MSAppCenter.setLogLevel(.verbose)
+        
+        // This method must only be used after Services have been started.
+        MSAppCenter.setTrackingEnabled(true)
+        
+    }
+}
+
+extension AppCenterOperation: MSDistributeDelegate {
+    func distribute(_ distribute: MSDistribute!,
+                    releaseAvailableWith details: MSReleaseDetails!) -> Bool {
+        
+        guard
+            let appDelegate = UIApplication.shared.delegate as? AppDelegate,
+            let window = appDelegate.window,
+            let rootViewController = appDelegate.appRootRouter?.rootViewController
+        else {
+            return false
+        }
+
+        let alertController = UIAlertController(title: "Update available \(details?.shortVersion ?? "") (\(details?.version ?? ""))",
+            message: "Release Note:\n\n\(details?.releaseNotes ?? "")\n\nDo you want to update?",
+            preferredStyle:.actionSheet)
+        alertController.configPopover(pointToView: window)
+
+        alertController.addAction(UIAlertAction(title: "Update", style: .cancel) {_ in
+            MSDistribute.notify(.update)
+        })
+
+        alertController.addAction(UIAlertAction(title: "Postpone", style: .default) {_ in
+            MSDistribute.notify(.postpone)
+        })
+
+        if let url = details.releaseNotesUrl {
+            alertController.addAction(UIAlertAction(title: "View release note", style: .default) {_ in
+                UIApplication.shared.open(url, options: [:])
+            })
+        }
+
+        alertController.addAction(UIAlertAction(title: "Cancel", style: .default) {_ in })
+
+        window.endEditing(true)
+        rootViewController.present(alertController, animated: true)
+
+        return true
+    }
+}
+
+extension AppCenterOperation: MSCrashesDelegate {
+
+    public func crashes(_ crashes: MSCrashes!, shouldProcessErrorReport errorReport: MSErrorReport!) -> Bool {
+        return !TrackingManager.shared.disableCrashSharing
+    }
+
+    public func crashes(_ crashes: MSCrashes!, didSucceedSending errorReport: MSErrorReport!) {
+        zmLog.error("AppCenter: finished sending the crash report")
+    }
+    
+    public func crashes(_ crashes: MSCrashes!, didFailSending errorReport: MSErrorReport!, withError error: Error!) {
+        zmLog.error("AppCenter: failed sending the crash report with error: \(error.localizedDescription)")
+    }
+}
